fix(condo): DOMA-11377 fixed styles for CookieAgreement
